Optimized some code for the stamper and large chemical tanks
